requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
groq
google-generativeai
pycountry
fastapi
uvicorn
